Add files via upload
This commit features a large improvement in usability - options have been added. The current list of options is now [-s], [-d], and [-h]. The first indicates the source directory from which the binaries can be obtained. The second indicates the destination directory (currently, mostly for use by the directory copying workflow). The third option is simply a help option, which will output a simple help statement. 

When the -s option is used and execution has reached the VirusTotal scanning portion, the user can choose to use the directory that they provided then, instead of having to re-type it.